week_7_8_9_project/requirements.txt
pandas
pyarrow
airflow
apache-airflow-providers-google
google-cloud-storage
